combined updates for core release, apphub, minibatch, msp edition, async rest apis, dependency updates   (#198)
integrates with omegaml-core 0.13.5
integrates with apphub minibatch
adds async rest apis.
enables command and per-user specified mongodb/db+credentials, rabbitmq/credentials+vhosts, effectively yielding a per-user celery cluster
separates omops from rest of cluster, improving operations security
enable operations APIs by landingpage, paasdeploy (sign up users via API)
enable kubespawner to accept user specific configuration updates

+++ scripts/omegajobs.sh
@@ -7,8 +7,8 @@
 ##    --singleuser     Run jupyterhub-singleuser
 ##    --ip=VALUE       ip address
 ##    --port=PORT      port
-##    --installonly    install only then exit
 ##    --debug          debug jupyterhub and notebook
+##    --label          runtime label
 ##
 ##    @script.name [option]
 
@@ -17,36 +17,56 @@ script_dir=$(dirname "$0")
 script_dir=$(realpath $script_dir)
 source $script_dir/easyoptions || exit
 
-ip=${ip:=0.0.0.0}
-port=${port:=5000}
-sitepackages=$(python -m site | grep site-packages | head -n 1 | cut -f 2 -d "'")
+# set defaults
+ip=${ip:-0.0.0.0}
+port=${port:-5000}
 omegaml_dir=$(python -W ignore -c  "import omegaml; print(omegaml.__path__[0])")
 omegajobs_dir=$(python -W ignore -c  "import omegajobs; print(omegajobs.__path__[0])")
+runtimelabel="${label:-$(hostname)},$CELERY_Q"
+
 
 if [[ ! -z $debug ]]; then
   jydebug="--debug"
 fi
 
-if [[ -z `which jupyterhub` || $installonly ]]; then
-  echo  "Installing jupyterhub"
-  conda install -y -c conda-forge jupyterhub=0.9.4 notebook=5.7.6
-  pip install -U jupyterhub-simplespawner==0.1 ipykernel==5.1.1 ipython==7.6.0
-  if [[ $installonly ]]; then
-    echo "Installation completed. Not starting at this time due to --installonly."
-    exit
-  fi
+export C_FORCE_ROOT=1
+export CELERY_Q=$runtimelabel
+# -- running in pod, use /app as a shared home
+if [[ -d "/app" ]]; then
+  export APPBASE="/app"
+  export PYTHONPATH="/app/pylib/user:/app/pylib/base"
+  export PYTHONUSERBASE="/app/pylib/user"
+  export OMEGA_CONFIG_FILE="app/pylib/user/.omegaml/config.yml"
+  export PATH="$PYTHONUSERBASE/bin:$PATH"
+else
+  export APPBASE=$HOME
+  export OMEGA_CONFIG_FILE="$APPBASE/.omegaml/config.yml"
 fi
 
 if [[ $singleuser ]]; then
     echo "Starting singleuser spawned juypter notebook"
-    mkdir -p $HOME/.jupyter
-    cp $omegaml_dir/notebook/jupyter/* $HOME/.jupyter/
+    if [[ ! -f $HOME/.jupyter/.omegaml ]]; then
+        mkdir -p $HOME/.jupyter
+        cp $omegaml_dir/notebook/jupyter/* $HOME/.jupyter/
+    fi
+    # -- if there is no config file, create one
+    if [[ ! -f $OMEGA_CONFIG_FILE ]]; then
+        mkdir -p $OMEGA_CONFIG_FILE
+        touch $OMEGA_CONFIG_FILE/config.yml
+    fi
+    pip install -U jupyterhub==$JY_HUB_VERSION jupyterlab
     cd $HOME/.jupyter
+    nohup honcho -d $APPBASE start worker >> worker.log 2>&1 &
+    jupyter serverextension enable jupyterlab
     jupyterhub-singleuser --ip $ip --port $port --allow-root $jydebug
 else
     echo "Starting multiuser spawned juypter hub"
-    # -- we only need & install kubespawner on omjobs, not in the spawned process
-    pip install -U jupyterhub-kubespawner==0.10.1
+    # make sure we have the proper env setup
+    # TODO move this to the runtime setup for the omegaml deployment
+    pip install postgres
+    image_pysite=$(find / -name omegajobs | grep site-packages)/..
+    cp -r $image_pysite/omegajobs /app/pylib/base
+    cp -r $image_pysite/omegaee /app/pylib/base
     CONFIGPROXY_AUTH_TOKEN=12345678 jupyterhub --ip $ip --port $port --config $omegajobs_dir/jupyterhub_config.py $jydebug
 fi
